added friendship controller
