+++ app/helpers/friendships_helper.rb
@@ -1,0 +1,2 @@
+module FriendshipsHelper
+end
