+++ config/routes.rb
@@ -2,13 +2,14 @@ Rails.application.routes.draw do
   devise_for :users, skip: :root
 
   resources :users, only: [:index, :show] do
+    resources :friendship, only: [:index]
     resources :posts, only: [:index, :create] do
       resources :likes, only: [:create]
       resources :comments, only: [:create]
     end    
   end
 
-  get 'users/:user_id/notify', to: 'users#notification', as: :friend_request_notify
+  #get 'users/:user_id/notify', to: 'users#notification', as: :friend_request_notify
   post 'friend', to: 'users#send_request', as: :send_friend_request
   patch 'friend', to: 'users#accept_request', as: :accept_friend_request
